Add a common checkpoint manager bridge
Summary:
Introduce a dependency-neutral checkpoint manager contract for TorchTitan and
group the checkpointing modules under `components/checkpointer/`.

`BaseCheckpointManager(Configurable, ABC)` owns the trainer-facing checkpoint
policy as its nested `Config`, the runtime contract as `@abstractmethod`s, the
checkpoint item names, and the model-state wrapper shared by concrete managers.
Keeping the config nested matches how every other Configurable works:
`Configurable.__init_subclass__` sets `Config._owner` on the class that declares
it and `build()` constructs `_owner`, so selecting a config selects its manager.
A config declared outside a manager could not be built at all.

Making the contract an abstract base rather than a `Protocol` means a manager
that omits a method fails at class definition instead of at call time, and
extension is the same shape as the rest of the repo: subclass, declare
`class Config(BaseCheckpointManager.Config)` with any extra fields, and `_owner`
auto-wires.

Module layout:

  components/checkpoint.py -> components/checkpointer/dcp_checkpointing_manager.py
  (new)                    -> components/checkpointer/base_checkpoint_manager.py
  (new)                    -> components/checkpointer/__init__.py

`checkpointer/__init__.py` re-exports the public names, so most call sites only
change the module path. `TorchCheckpointingManager` is excluded: it imports the
optional `torch_checkpointing` dependency at module scope, so re-exporting it
would make every importer of the package require it.

Keep the existing DCP `CheckpointManager` as the default implementation while its
config defines only the DCP-specific `async_mode`; all shared fields and
validation live in the base config. Type core Trainer, Forge, and the RL
checkpoint build sites against the neutral contract so an independent manager can
be selected without adding backend branches to the legacy manager. TorchFT remains
explicitly tied to the legacy DCP implementation.

Test Plan:
The unit suite cannot be collected in this environment: the available torch
predates `torch.distributed.fsdp.DataParallelMeshDims`, which current main imports,
so collection fails on unmodified main too.

Verified directly, with `spmd_types` installed:
- `python -m compileall torchtitan tests scripts` is clean.
- `flake8 --config=.flake8` and `ufmt check` are clean on the changed files, and
  `codespell` reports nothing.
- `pyrefly check` in project mode reports 1 error, the same one `main` reports.
- No references remain to `components.checkpoint`, `checkpoint_utils`,
  `CheckpointManagerProtocol`, or `CheckpointManagerConfig`.
- `torchtitan.components.optimizer`, `torchtitan.components.optimizer_utils`, and
  `torchtitan.components.checkpointer` each import cleanly on their own,
  confirming no import cycle.
- `BaseCheckpointManager.__abstractmethods__` is
  `{load, save, maybe_wait_for_staging, close}`; the base cannot be instantiated.
- `CheckpointManager` subclasses the base and `Config._owner` resolves to it;
  `async_mode` stays on the DCP config only.
- A throwaway subclass with an extra config field builds via `Config(...).build()`
  and returns that subclass; inherited validation still raises on
  `keep_latest_k=1`.

CI is expected to cover the suite.
