+++ torchtitan/components/checkpointer/__init__.py
@@ -1,0 +1,35 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
+from .base_checkpoint_manager import (
+    BaseCheckpointManager,
+    DATALOADER,
+    LR_SCHEDULER,
+    MODEL,
+    ModelWrapper,
+    OPTIMIZER,
+    purge_worker,
+    TRAIN_STATE,
+)
+from .dcp_checkpointing_manager import AsyncMode, CheckpointManager
+
+# ``TorchCheckpointingManager`` is deliberately not re-exported. It imports
+# ``torch_checkpointing`` at module scope, and that is an optional dependency, so
+# re-exporting it would make every importer of this package require it. Import it
+# from ``torchtitan.components.checkpointer.torch_checkpointing_manager``.
+
+__all__ = [
+    "AsyncMode",
+    "BaseCheckpointManager",
+    "CheckpointManager",
+    "DATALOADER",
+    "LR_SCHEDULER",
+    "MODEL",
+    "ModelWrapper",
+    "OPTIMIZER",
+    "TRAIN_STATE",
+    "purge_worker",
+]
