+++ torchtitan/experiments/rl/examples/dapo_math/config_registry.py
@@ -8,7 +8,7 @@
 
 from __future__ import annotations
 
-from torchtitan.components.checkpoint import CheckpointManager
+from torchtitan.components.checkpointer import CheckpointManager
 from torchtitan.components.loss import ChunkedLossWrapper
 from torchtitan.components.lr_scheduler import LRSchedulersContainer
 from torchtitan.components.optimizer import default_adamw
