+++ torchtitan/experiments/torchft/tests/test_torchft_checkpoint.py
@@ -111,7 +111,7 @@ class TestFTCheckpointManager(unittest.TestCase):
 
     @mock.patch("torch.cuda.Stream")
     @mock.patch(
-        "torchtitan.components.checkpoint.dcp.async_save", side_effect=fake_async_save
+        "torchtitan.components.checkpointer.dcp_checkpointing_manager.async_save", side_effect=fake_async_save
     )
     def test_torchft_async_save_calls_maybe_wait_for_saving(
         self,
